+++ src/homework1/task15.py
@@ -1,16 +1,26 @@
-# 15 lines: itertools
-from itertools import groupby
-lines = '''
-This is the
-first paragraph.
+# 16 lines: csv module, tuple unpacking, cmp() built-in
+import csv
 
-This is the second.
-'''.splitlines()
-# Use itertools.groupby and bool to return groups of
-# consecutive lines that either have content or don't.
-for has_chars, frags in groupby(lines, bool):
-    if has_chars:
-        print(' '.join(frags))
-# PRINTS:
-# This is the first paragraph.
-# This is the second.
+
+# need to define cmp function in Python 3
+def cmp(a, b):
+    return (a > b) - (a < b)
+
+
+# write stocks data as comma-separated values
+with open('stocks.csv', 'w', newline='') as stocksFileW:
+    writer = csv.writer(stocksFileW)
+    writer.writerows([
+        ['GOOG', 'Google, Inc.', 505.24, 0.47, 0.09],
+        ['YHOO', 'Yahoo! Inc.', 27.38, 0.33, 1.22],
+        ['CNET', 'CNET Networks, Inc.', 8.62, -0.13, -1.4901]
+    ])
+
+# read stocks data, print status messages
+with open('stocks.csv', 'r') as stocksFile:
+    stocks = csv.reader(stocksFile)
+
+    status_labels = {-1: 'down', 0: 'unchanged', 1: 'up'}
+    for ticker, name, price, change, pct in stocks:
+        status = status_labels[cmp(float(change), 0.0)]
+        print('%s is %s (%.2f)' % (name, status, float(pct)))
